null check

+++ src/components/blocks/audio/Audio.vue
@@ -152,7 +152,7 @@ export default {
             this.upload.completed = false;
           }
         } catch (error) {
-          if ([415, 413].includes(error.response.status)) {
+          if (error.response && [415, 413].includes(error.response.status)) {
             this.deleteNode();
           } else {
             const editorVm = this.getEditorVm();
